requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
distro==1.9.0
Flask==3.1.0
frozendict==2.4.6
frozenlist==1.5.0
future==1.0.0
googlesearch-python==1.2.5
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
line-bot-sdk==3.14.2
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.1
openai==1.58.1
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
yfinance==0.2.51
